+++ RestSharp.Portable.Core/RestClientBase.cs
@@ -25,11 +25,11 @@ namespace RestSharp.Portable
 
         private readonly IList<string> _acceptEncodings = new List<string>();
 
-        private readonly List<Parameter> _defaultParameters = new List<Parameter>();
+        private readonly IParameterCollection _defaultParameters = new ParameterCollection();
+
+        private readonly Lazy<IHttpClient> _httpClient;
 
         private bool _disposedValue; // For the detection of multiple calls
-
-        private IHttpClient _httpClient;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="RestClientBase" /> class.
@@ -37,6 +37,8 @@ namespace RestSharp.Portable
         /// <param name="httpClientFactory">The HTTP client factory to use</param>
         protected RestClientBase(IHttpClientFactory httpClientFactory)
         {
+            _httpClient = new Lazy<IHttpClient>(() => httpClientFactory.CreateClient(this));
+
             HttpClientFactory = httpClientFactory;
 
             var jsonDeserializer = new JsonDeserializer();
@@ -110,14 +112,6 @@ namespace RestSharp.Portable
         public CookieContainer CookieContainer { get; set; }
 
         /// <summary>
-        /// Gets or sets the default <see cref="StringComparer"/> to be used for the requests.
-        /// </summary>
-        /// <remarks>
-        /// If this property is null, the <see cref="StringComparer.Ordinal"/> is used.
-        /// </remarks>
-        public StringComparer DefaultParameterNameComparer { get; set; }
-
-        /// <summary>
         /// Gets or sets the timeout to be used for requests.
         /// </summary>
         /// <remarks>
@@ -136,34 +130,19 @@ namespace RestSharp.Portable
         {
             get
             {
-                var comparer = DefaultParameterNameComparer ?? StringComparer.OrdinalIgnoreCase;
-                return (string)DefaultParameters.FirstOrDefault(x => comparer.Equals(x.Name, "User-Agent"))?.Value;
+                var userAgentParameter = DefaultParameters.Find(ParameterType.HttpHeader, "User-Agent").FirstOrDefault();
+                return (string)userAgentParameter?.Value;
             }
             set
             {
-                var comparer = DefaultParameterNameComparer ?? StringComparer.OrdinalIgnoreCase;
-                var parameter = DefaultParameters.FirstOrDefault(x => comparer.Equals(x.Name, "User-Agent"));
-                if (parameter == null)
-                {
-                    parameter = new Parameter()
-                    {
-                        Name = "User-Agent",
-                        Type = ParameterType.HttpHeader,
-                        ValidateOnAdd = false,
-                    };
-                    DefaultParameters.Add(parameter);
-                }
-                parameter.Value = value;
+                DefaultParameters.AddOrUpdate(new Parameter { Type = ParameterType.HttpHeader, Name = "User-Agent", Value = value, ValidateOnAdd = false });
             }
         }
 
         /// <summary>
         /// Gets the collection of the default parameters for all requests
         /// </summary>
-        public IList<Parameter> DefaultParameters
-        {
-            get { return _defaultParameters; }
-        }
+        public IParameterCollection DefaultParameters => _defaultParameters;
 
         /// <summary>
         /// Gets or sets the credentials used for the request (e.g. NTLM authentication)
@@ -436,8 +415,9 @@ namespace RestSharp.Portable
         /// Gets the content for a request.
         /// </summary>
         /// <param name="request">The <see cref="IRestRequest"/> to get the content for.</param>
+        /// <param name="parameters">The request parameters for the REST request (read-only)</param>
         /// <returns>The <see cref="IHttpContent"/> for the <paramref name="request"/></returns>
-        protected abstract IHttpContent GetContent(IRestRequest request);
+        protected abstract IHttpContent GetContent(IRestRequest request, RequestParameters parameters);
 
         /// <summary>
         /// Allows the implementor to modify the <paramref name="httpClient"/> and the <paramref name="requestMessage"/>
@@ -457,7 +437,6 @@ namespace RestSharp.Portable
         /// <returns>The <see cref="IHttpResponseMessage"/> for the request</returns>
         protected async Task<IHttpResponseMessage> ExecuteRequest(IRestRequest request, CancellationToken ct)
         {
-            AddDefaultParameters(request);
             while (true)
             {
                 if (Authenticator != null && Authenticator.CanPreAuthenticate(this, request, Credentials))
@@ -465,41 +444,15 @@ namespace RestSharp.Portable
                     await Authenticator.PreAuthenticate(this, request, Credentials);
                 }
 
-                // Lazy initialization of the HTTP client
-                if (_httpClient == null)
-                {
-                    _httpClient = HttpClientFactory.CreateClient(this, request);
-                }
-
+                var requestParameters = this.MergeParameters(request);
                 bool failed = true;
-                var message = HttpClientFactory.CreateRequestMessage(this, request);
+                var message = HttpClientFactory.CreateRequestMessage(this, request, requestParameters.OtherParameters);
                 try
                 {
-                    var bodyData = GetContent(request);
+                    var bodyData = GetContent(request, requestParameters);
                     if (bodyData != null)
                     {
                         message.Content = bodyData;
-                    }
-
-                    if (message.Content != null)
-                    {
-                        var content = message.Content;
-                        foreach (var param in request.Parameters.Where(x => x.Type == ParameterType.HttpHeader && x.IsContentParameter()))
-                        {
-                            if (content.Headers.Contains(param.Name))
-                            {
-                                content.Headers.Remove(param.Name);
-                            }
-
-                            if (param.ValidateOnAdd)
-                            {
-                                content.Headers.Add(param.Name, param.ToRequestString());
-                            }
-                            else
-                            {
-                                content.Headers.TryAddWithoutValidation(param.Name, param.ToRequestString());
-                            }
-                        }
                     }
 
                     ModifyRequestBeforeAuthentication(_httpClient, message);
@@ -602,23 +555,6 @@ namespace RestSharp.Portable
         }
 
         /// <summary>
-        /// Add overridable default parameters to the request
-        /// </summary>
-        /// <param name="request">The requests to add the default parameters to.</param>
-        private void AddDefaultParameters(IRestRequest request)
-        {
-            var comparer = new ParameterComparer(this, request);
-
-            var startIndex = 0;
-            foreach (var parameter in DefaultParameters.Where(x => x.Type != ParameterType.HttpHeader))
-            {
-                if (request.Parameters.Contains(parameter, comparer))
-                {
-                    continue;
-                }
-
-                request.Parameters.Insert(startIndex++, parameter);
-            }
         }
     }
 }